api/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
clear